refactor(cache): move publication rules to kv_cache/publish.py
Split paged_cache_spec.py along the contract/publisher boundary:

- configs/paged_cache_spec.py keeps the contract types and scheduler-side
  logic shared by configs, engine, execution and backends:
  PagedCacheGroupSpec, Retention/Family, the layer_type label vocabulary,
  compute_paged_cache_group_page_counts, validate_scheduler_config and
  compute_max_logical_pages_for_capture.
- The KV-pool-side publication rules move to
  layers/attention/kv_cache/publish.py: publish_paged_cache_groups,
  group_specs_from_layer_types, layer_group_ids,
  split_recurrent_state_groups and hybrid_slab_group_size, plus their
  private label-retention helpers.

This straightens the dependency direction: model config files no longer
share a module with pool publication logic, and kv_cache (the recipes'
single source of group ids) now owns the derivation it documents.

File-path-loading tests register the contract module under its real name
so publish.py's from-import binds the repo file via the sys.modules
short-circuit. The legacy-layout sitecustomize now patches only the
publish module, whose sole consumer binds at call time.
