+++ python/tokenspeed/runtime/layers/attention/kv_cache/msa.py
@@ -33,6 +33,7 @@ class MSATokenToKVPool(MHATokenToKVPool):
         indexed_layer_ids: frozenset[int],
         memory_plan: CacheMemoryPlan,
         layer_types: tuple[str, ...] = (),
+        layer_group_ids: tuple[str, ...] = (),
         sliding_window_tokens: int | tuple[int | None, ...] | None = None,
         max_scheduled_tokens: int = 0,
         pd_disaggregation_enabled: bool = False,
@@ -54,6 +55,7 @@ class MSATokenToKVPool(MHATokenToKVPool):
             page_size=page_size,
             rank=rank,
             layer_types=layer_types,
+            layer_group_ids=layer_group_ids,
             sliding_window_tokens=sliding_window_tokens,
             max_scheduled_tokens=max_scheduled_tokens,
             pd_disaggregation_enabled=pd_disaggregation_enabled,
