+++ test/runtime/models/test_inkling_models.py
@@ -118,7 +118,7 @@ class TestInklingConfigRegistry(unittest.TestCase):
         """Cache-group labels: sliding layers split round-robin into
         equal-count sub-groups sized by the full-layer count, so every
         hybrid slab is bound by one layer of each group (5+1 truncated,
-        5x11+11 at full depth; see paged_cache_spec.hybrid_slab_group_size).
+        5x11+11 at full depth; see kv_cache.publish.hybrid_slab_group_size).
         Exposed as paged_cache_layer_types, NOT layer_types: transformers
         validates layer_types against ALLOWED_LAYER_TYPES, which rejects
         sub-group labels."""
