+++ test/runtime/test_host_executor.py
@@ -164,6 +164,7 @@ class MemoryExecutorTest(unittest.TestCase):
     def setUp(self):
         try:
             import torch
+            from cache_pool_test_utils import plan_fields
             from tokenspeed_scheduler import Cache
 
             from tokenspeed.runtime.cache.executor.memory_executor import (
@@ -175,9 +176,6 @@ class MemoryExecutorTest(unittest.TestCase):
             )
             from tokenspeed.runtime.layers.attention.kv_cache.mha import (
                 MHATokenToKVPool,
-            )
-            from tokenspeed.runtime.layers.attention.kv_cache.plan import (
-                plan_cache_fields,
             )
             from tokenspeed.runtime.layers.attention.kv_cache.recipes.qwen35 import (
                 qwen_gdn_cache_fields,
@@ -195,7 +193,7 @@ class MemoryExecutorTest(unittest.TestCase):
         self.MHATokenToKVPool = MHATokenToKVPool
         self.HybridMHATokenToKVPool = HybridMHATokenToKVPool
         self.qwen_gdn_cache_fields = qwen_gdn_cache_fields
-        self.plan_cache_fields = plan_cache_fields
+        self.plan_fields = plan_fields
 
     def _pool(self):
         from cache_pool_test_utils import make_mha_memory_plan
@@ -215,6 +213,8 @@ class MemoryExecutorTest(unittest.TestCase):
             "layer_types": LAYER_TYPES,
             "sliding_window_tokens": 128,
         }
+        from cache_pool_test_utils import make_layer_group_ids
+
         kwargs["memory_plan"] = make_mha_memory_plan(
             size=kwargs["size"],
             page_size=kwargs["page_size"],
@@ -222,6 +222,11 @@ class MemoryExecutorTest(unittest.TestCase):
             kv_heads=kwargs["head_num"],
             head_dim=kwargs["head_dim"],
             dtype=kwargs["dtype"],
+            layer_types=kwargs["layer_types"],
+            sliding_window_tokens=kwargs["sliding_window_tokens"],
+        )
+        kwargs["layer_group_ids"] = make_layer_group_ids(
+            layer_num=kwargs["layer_num"],
             layer_types=kwargs["layer_types"],
             sliding_window_tokens=kwargs["sliding_window_tokens"],
         )
@@ -335,7 +340,7 @@ class MemoryExecutorTest(unittest.TestCase):
             ssm_shape=(2, 8),
             ssm_element_size=2,
         )
-        plan = self.plan_cache_fields(
+        plan = self.plan_fields(
             fields,
             logical_block_tokens=4,
             num_lcm_blocks=4,
