+++ test/runtime/test_inkling_activation_parity.py
@@ -176,6 +176,7 @@ class _Harness:
             max_context_len=1024,
             page_size=PAGE_SIZE,
             rank=0,
+            layer_group_ids=("full_attention",) * text.num_hidden_layers,
             memory_plan=make_mha_memory_plan(
                 size=1024,
                 page_size=PAGE_SIZE,
