+++ test/runtime/test_mxfp8_kv_pool.py
@@ -58,6 +58,7 @@ def _make_pool(page_size: int, size: int = 512):
         max_context_len=size,
         page_size=page_size,
         rank=0,
+        layer_group_ids=("full_attention",) * LAYERS,
         memory_plan=make_mha_memory_plan(
             size=size,
             page_size=page_size,
@@ -187,6 +188,7 @@ def _make_shared_pool(size: int = 512):
         page_size=128,
         rank=0,
         layer_types=SHARED_LAYER_TYPES,
+        layer_group_ids=SHARED_LAYER_TYPES,
         sliding_window_tokens=512,
         max_scheduled_tokens=size,
         slot_tokens=256,
